pyproject.toml
[project]
name = "arxivwatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Major Hayden", email = "major@mhtx.net" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "email-validator>=2.3.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "markdown>=3.9",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.1.1",
    "structlog>=25.4.0",
]

[project.scripts]
arxivwatch = "arxivwatch:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
]
